+++ ESFramework/Core/ESHelpers.h
@@ -87,11 +87,6 @@ FOUNDATION_EXTERN BOOL ESOSVersionIsAtLeast(NSInteger majorVersion, NSInteger mi
 FOUNDATION_EXTERN void ESMeasureExecution(NS_NOESCAPE void (^block)(void), NS_NOESCAPE void (^ _Nullable completion)(NSTimeInterval elapsedMillisecond));
 
 /**
- * Generates an UUID string, e.g. "E621E1F8-C36C-495A-93FC-0C247A3E6E5F".
- */
-FOUNDATION_EXTERN NSString *ESUUIDString(void);
-
-/**
  * Generates a random number between min and max.
  */
 FOUNDATION_EXTERN uint32_t ESRandomNumber(uint32_t min, uint32_t max);