+++ ESFramework/Foundation/NSString+ESExtension.h
@@ -13,6 +13,11 @@ NS_ASSUME_NONNULL_BEGIN
 @interface NSString (ESExtension)
 
 /**
+ * Generates an UUID string, e.g. "E621E1F8-C36C-495A-93FC-0C247A3E6E5F".
+ */
++ (NSString *)UUIDString;
+
+/**
  * Returns a NSNumber object created by parsing the string.
  */
 - (nullable NSNumber *)numberValue;